# server/python/src/mfs_server/api/app.py
"""FastAPI /v1 control plane (design/02 §1, 03). Thin HTTP wrappers over Engine.
Typed request/response models (api/models.py) make the generated OpenAPI rich enough
for the multi-language SDKs. `add` indexes inline by default (returns job_id when done)
or enqueues for the standalone worker when process=false.
"""
from __future__ import annotations

from contextlib import asynccontextmanager

from fastapi import FastAPI, HTTPException, Request
from fastapi.exceptions import RequestValidationError
from fastapi.responses import JSONResponse

from ..config import ServerConfig, load_server_config
from ..engine.engine import Engine
from .models import (
    AddRequest, AddResponse, CancelResponse, CatMeta, CatResponse, EstimateResponse,
    GrepResponse, JobResponse, LsResponse, ManifestRequest, ManifestResponse,
    ProbeRequest, ProbeResponse, RemoveResponse, SearchResponse, ServerInfo, StatusResponse,
)

# Canonical error codes -> suggested next actions (protocol/errors.md). The endpoints
# raise HTTPException with the canonical code as `detail` for these cases; the handler
# below turns that into the stable {code, detail, suggestions} envelope SDKs switch on.
_CODE_SUGGESTIONS = {
    "object_too_large_for_cat": ["head", "cat --range", "export"],
    "is_directory": ["ls", "tree"],
    "range_unsupported": ["cat --meta", "export"],
    "density_unsupported": ["head", "cat --range"],
    "tail_unsupported": ["head", "cat --range"],
    "locator_not_found": ["re-search; the record may have changed"],
    "since_unsupported": ["drop --since"],
    "sync_already_running": ["status <uri>", "job cancel <id>"],
    "connector_removing": ["wait for removal to finish, then retry"],
    "connector_unhealthy": ["check credentials/connectivity"],
    "not_found": ["check the URI"],
}
# HTTP status -> code when `detail` isn't already a canonical code (human strings).
_STATUS_CODE = {400: "bad_request", 404: "not_found", 409: "conflict",
                422: "validation_error", 502: "connector_unhealthy"}


def create_app(cfg: ServerConfig | None = None) -> FastAPI:
    cfg = cfg or load_server_config()

    @asynccontextmanager
    async def lifespan(app: FastAPI):
        eng = Engine(cfg)
        await eng.startup()
        app.state.engine = eng
        try:
            yield
        finally:
            await eng.shutdown()

    app = FastAPI(title="MFS", version="0.4.0", lifespan=lifespan,
                  description="Multi-source File-like Search — HTTP /v1 control plane.")

    if cfg.auth_token:
        @app.middleware("http")
        async def _auth(request: Request, call_next):
            """Bearer-token gate (design/02 §5.3 / §11.2): when auth_token is configured,
            every request — loopback included — must carry Authorization: Bearer <token>."""
            if request.headers.get("authorization", "") != f"Bearer {cfg.auth_token}":
                return JSONResponse(status_code=401, content={
                    "code": "unauthorized", "detail": "missing or invalid bearer token",
                    "suggestions": ["set a profile token (Authorization: Bearer <token>)"]})
            return await call_next(request)

    @app.exception_handler(HTTPException)
    async def _http_exc(_request: Request, exc: HTTPException) -> JSONResponse:
        """Wrap HTTPException into the {code, detail, suggestions} envelope (errors.md).
        When `detail` is already a canonical code, surface it as `code`; otherwise derive
        `code` from the HTTP status and keep the human string as `detail`."""
        detail = exc.detail if isinstance(exc.detail, str) else "error"
        code = detail if detail in _CODE_SUGGESTIONS else _STATUS_CODE.get(exc.status_code, "error")
        return JSONResponse(
            status_code=exc.status_code,
            content={"code": code, "detail": detail, "suggestions": _CODE_SUGGESTIONS.get(code, [])})

    @app.exception_handler(RequestValidationError)
    async def _val_exc(_request: Request, exc: RequestValidationError) -> JSONResponse:
        return JSONResponse(status_code=422, content={
            "code": "validation_error", "detail": str(exc), "suggestions": ["fix request shape"]})

    @app.exception_handler(Exception)
    async def _unhandled_exc(_request: Request, exc: Exception) -> JSONResponse:
        """Any uncaught error still returns the stable envelope so SDKs can switch on
        `code` instead of parsing a raw 500 body."""
        return JSONResponse(status_code=500, content={
            "code": "internal_error", "detail": str(exc), "suggestions": []})

    def eng() -> Engine:
        return app.state.engine

    @app.get("/v1/server/info", response_model=ServerInfo, operation_id="getServerInfo", tags=["server"])
    async def server_info() -> ServerInfo:
        import socket
        return ServerInfo(version="0.4.0", machine_id=socket.gethostname(), namespace=cfg.namespace)

    @app.post("/v1/add", response_model=AddResponse, operation_id="addSource", tags=["ingest"])
    async def add(body: AddRequest) -> AddResponse:
        try:
            job_id = await eng().add(body.target, config=body.config, full=body.full,
                                     since=body.since, process=body.process, update_config=body.update)
        except ValueError as e:
            code = str(e)
            status = 409 if code in ("sync_already_running", "connector_removing") else 400
            raise HTTPException(status, code)        # -> error envelope
        return AddResponse(job_id=job_id)

    @app.post("/v1/jobs/{job_id}/cancel", response_model=CancelResponse,
              operation_id="cancelJob", tags=["ingest"])
    async def cancel_job(job_id: str) -> CancelResponse:
        ok = await eng().cancel_job(job_id)
        return CancelResponse(job_id=job_id, cancelled=ok)

    @app.post("/v1/connectors/probe", response_model=ProbeResponse,
              operation_id="probeConnector", tags=["connectors"])
    async def probe(body: ProbeRequest) -> ProbeResponse:
        return ProbeResponse(**await eng().probe(body.target, body.config))

    @app.post("/v1/connectors/estimate", response_model=EstimateResponse,
              operation_id="estimateConnector", tags=["connectors"])
    async def estimate(body: ProbeRequest) -> EstimateResponse:
        """Zero-billing pre-flight estimate (design/04 §3): object/chunk/token counts via
        metadata + a local chunker/tokenizer dry-run. No embedding API calls."""
        return EstimateResponse(**await eng().estimate(body.target, body.config))

    @app.get("/v1/connectors/inspect", operation_id="inspectConnector", tags=["connectors"])
    async def inspect(target: str):
        out = await eng().inspect(target)
        if out is None:
            raise HTTPException(404, "connector not found")
        return out

    @app.delete("/v1/connectors", response_model=RemoveResponse,
                operation_id="removeConnector", tags=["connectors"])
    async def remove(target: str) -> RemoveResponse:
        return RemoveResponse(target=target, removed=await eng().remove_connector(target))

    @app.post("/v1/upload", response_model=AddResponse, operation_id="uploadSource", tags=["ingest"])
    async def upload(request: Request, name: str, process: bool = True) -> AddResponse:
        """CS upload flow: POST a tar(.gz) of a tree as the raw body (?name=<label>);
        the server stages + indexes it. For client/server without a shared filesystem."""
        data = await request.body()
        if not data:
            raise HTTPException(400, "empty upload body")
        try:
            out = await eng().ingest_upload(name, data, process=process)
        except ValueError as e:
            raise HTTPException(400, str(e))
        return AddResponse(job_id=out["job_id"])

    @app.post("/v1/files/manifest", response_model=ManifestResponse,
              operation_id="filesManifest", tags=["ingest"])
    async def files_manifest(body: ManifestRequest) -> ManifestResponse:
        """Manifest-diff upload step ②: stat-only manifest in, need_sha1 + deletion
        candidates out (design/02 §4.2). No bytes transferred here."""
        out = await eng().files_manifest(body.client_id, body.name,
                                         [f.model_dump() for f in body.files])
        return ManifestResponse(**out)

    @app.put("/v1/files/upload", response_model=AddResponse,
             operation_id="filesUpload", tags=["ingest"])
    async def files_upload(request: Request, client_id: str, name: str,
                           process: bool = True) -> AddResponse:
        """Manifest-diff upload step ④: PUT a tar(.gz) carrying a `.mfs-meta.json`
        member (hashes/renames/deletions) + the changed file bytes. The server applies
        it to the staging area and triggers the file-connector sync."""
        data = await request.body()
        if not data:
            raise HTTPException(400, "empty upload body")
        try:
            out = await eng().files_upload(client_id, name, data, process=process)
        except ValueError as e:
            raise HTTPException(400, str(e))
        return AddResponse(job_id=out["job_id"])

    @app.get("/v1/search", response_model=SearchResponse, operation_id="search", tags=["retrieval"])
    async def search(q: str, path: str | None = None, mode: str = "hybrid",
                     top_k: int = 10, collapse: bool = False) -> SearchResponse:
        connector_uri = None
        object_prefix = None
        if path:
            connector_uri, object_prefix = await eng().resolve_connector_uri(path)
        results = await eng().search(q, connector_uri=connector_uri, object_prefix=object_prefix,
                                     mode=mode, top_k=top_k, collapse=collapse)
        return SearchResponse(results=results)

    @app.get("/v1/grep", response_model=GrepResponse, operation_id="grep", tags=["retrieval"])
    async def grep(pattern: str, path: str) -> GrepResponse:
        return GrepResponse(results=await eng().grep(pattern, path))

    @app.get("/v1/ls", response_model=LsResponse, operation_id="ls", tags=["browse"])
    async def ls(path: str) -> LsResponse:
        try:
            return LsResponse(**await eng().ls(path))
        except (FileNotFoundError, NotADirectoryError, ValueError) as e:
            raise HTTPException(404, str(e))

    @app.get("/v1/cat", operation_id="cat", tags=["browse"],
             response_model=None, responses={200: {"model": CatResponse}})
    async def cat(path: str, range: str | None = None, meta: bool = False,
                  density: str | None = None, locator: str | None = None):
        import json as _json
        rg = None
        if range:
            a, b = range.split(":")
            rg = (int(a), int(b))
        loc = None
        if locator:
            try:
                loc = _json.loads(locator)
            except ValueError:
                raise HTTPException(400, "invalid locator JSON")
        try:
            out = await eng().cat(path, range=rg, meta=meta, density=density, locator=loc)
        except IsADirectoryError:
            raise HTTPException(400, "is_directory")
        except ValueError as e:
            code = str(e)
            if code in ("density_unsupported", "range_unsupported", "object_too_large_for_cat"):
                raise HTTPException(400, code)
            if code == "locator_not_found":
                raise HTTPException(404, "locator_not_found")
            raise HTTPException(404, code)
        except FileNotFoundError as e:
            raise HTTPException(404, str(e))
        if meta:
            return CatMeta(**out) if isinstance(out, dict) else out
        if isinstance(out, dict):     # locator hit -> {source, locator, content}
            return CatResponse(source=out.get("source", path), content=out.get("content", ""))
        return CatResponse(source=path, content=out)

    async def _read_op(fn, path: str):
        """Shared error mapping for head/tail/export."""
        try:
            return await fn(path)
        except IsADirectoryError:
            raise HTTPException(400, "is_directory")
        except FileNotFoundError as e:
            raise HTTPException(404, str(e))
        except ValueError as e:
            raise HTTPException(400, str(e))

    @app.get("/v1/head", response_model=CatResponse, operation_id="head", tags=["browse"])
    async def head(path: str, n: int = 20) -> CatResponse:
        return CatResponse(source=path, content=await _read_op(lambda p: eng().head(p, n), path))

    @app.get("/v1/tail", response_model=CatResponse, operation_id="tail", tags=["browse"])
    async def tail(path: str, n: int = 20) -> CatResponse:
        return CatResponse(source=path, content=await _read_op(lambda p: eng().tail(p, n), path))

    @app.get("/v1/export", response_model=CatResponse, operation_id="export", tags=["browse"])
    async def export(path: str) -> CatResponse:
        """Full object content (no row cap / size guard) for `mfs export`."""
        return CatResponse(source=path, content=await _read_op(eng().export, path))

    @app.get("/v1/status", response_model=StatusResponse, operation_id="status", tags=["server"])
    async def status() -> StatusResponse:
        conns = await eng().meta.fetchall(
            "SELECT root_uri, type, status FROM connectors WHERE namespace_id=?", (cfg.namespace,))
        jobs = await eng().meta.fetchall(
            "SELECT status, count(*) AS n FROM connector_jobs GROUP BY status")
        return StatusResponse(connectors=[dict(c) for c in conns],
                              jobs={j["status"]: j["n"] for j in jobs})

    @app.get("/v1/jobs", response_model=list[JobResponse], operation_id="listJobs", tags=["ingest"])
    async def list_jobs(limit: int = 20) -> list[JobResponse]:
        rows = await eng().meta.fetchall(
            "SELECT * FROM connector_jobs ORDER BY started_at DESC LIMIT ?", (limit,))
        return [JobResponse(**{k: dict(r).get(k) for k in JobResponse.model_fields}) for r in rows]

    @app.get("/v1/jobs/{job_id}", response_model=JobResponse, operation_id="getJob", tags=["ingest"])
    async def job(job_id: str) -> JobResponse:
        row = await eng().meta.fetchone("SELECT * FROM connector_jobs WHERE id=?", (job_id,))
        if not row:
            raise HTTPException(404, "job not found")
        return JobResponse(**{k: dict(row).get(k) for k in JobResponse.model_fields})

    return app